churn_prediction_app/churn_prediction_app/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
xgboost>=1.5.0
imbalanced-learn>=0.8.0
requests>=2.27.0
python-dotenv>=0.20.0